ui/rbac-poc-actions-menu: move actions menu separately
