+++ ui/apps/everest/src/components/table-actions-menu/index.ts
@@ -1,0 +1,1 @@
+export { default } from './table-actions-menu';
