+++ ui/apps/everest/src/components/table-actions-menu/table-actions-menu.types.ts
@@ -1,0 +1,10 @@
+import { ReactNode } from 'react';
+import { ButtonProps, MenuProps } from '@mui/material';
+
+export interface TableActionsMenuProps {
+  menuItems: ReactNode[];
+  buttonProps?: ButtonProps;
+  menuProps?: Omit<MenuProps, 'open'> & {
+    showMenu?: boolean;
+  };
+}
